requirements.txt
torch=2.1.0+cu118
torchaudio=2.1.0+cu118
torchvision=0.16.0+cu118
Pillow=10.1.0
pandas=2.0.3
numpy=1.24.3
ipykernel=6.26.0
Django=3.2.20
